refactor(adapters): move gh from JS spawn to SKILL pipeline
- Degrade find-existing-task adapter to a stdin to JSON parser; drop cross-spawn and `--issue`/`--repo` flags (cross-spawn still used by platform-sync.js)
- Move POSIX/Windows gh pipeline into issue-pr-commands.github.{en,zh-CN}.md; import-issue SKILL references the rule and stays platform-agnostic
- Merge platform-adapter variants: parser is now platform-agnostic; .github.js variant deleted, base slot handles all platforms
- Rewrite tests to pipe JSON via stdin; add empty-stdin and invalid-JSON cases (8 tests total)
